Rename mappers package to ext, as it may contain other JAX-RS extensions, not only ExceptionMappers
